refactor(json-schema-annotation): make annotator process cleaner
