+++ core/shared/json-schema-annotation/readme.adoc
@@ -1,6 +1,5 @@
 == Add additional annotations like swagger to the jsonschema ==
 
-TIP: includeAdditionalProperties will be used as class annotation.
 
 == Supported annotations ==
 
@@ -93,14 +92,14 @@ source schema file
 {
   "title": "Person",
   "type": "object",
-  "swagger-schema": {
+  "swagger": {
     "name": "Person examples",
     "description": "schema for Person examples"
   },
   "properties": {
     "name": {
       "type": "string",
-      "schema": {
+      "swagger": {
         "type": "String",
         "description": "The schema of the person",
         "example": "John Doe",
